Set Folder path (#15)
폴더 경로를 재설정 한다.

Fixes https://github.com/dbwjd/BBKK/issues/14
