+++ app/src/main/java/com/bbkk/android/bbkkclient/view/main/MainActivity.java
@@ -1,7 +1,9 @@
-package com.bbkk.android.bbkkclient;
+package com.bbkk.android.bbkkclient.view.main;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.bbkk.android.bbkkclient.R;
 
 public class MainActivity extends AppCompatActivity {
