+++ app/src/main/res/layout/item_start.xml
@@ -2,21 +2,21 @@
 <android.support.constraint.ConstraintLayout
   xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent">
 
   <ImageButton
     android:id="@+id/btn_start_next"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
+    android:layout_width="40dp"
+    android:layout_height="40dp"
     android:layout_marginEnd="8dp"
     android:layout_marginStart="8dp"
-    android:layout_marginTop="4dp"
+    android:layout_marginTop="30dp"
     app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintHorizontal_bias="0.877"
     app:layout_constraintStart_toEndOf="@+id/textView_start_message"
     app:layout_constraintTop_toTopOf="parent"
-    app:srcCompat="@android:drawable/alert_dark_frame" />
+    app:srcCompat="@drawable/ic_launcher_foreground" />
 
   <TextView
     android:id="@+id/textView_start_message"
