+++ app/src/main/res/layout/item_tendency_view_flipper1.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout
+  xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
-  android:background="#f00"
+  android:background="#ecdec7"
   android:orientation="vertical"
-  android:layout_margin="20dp"
   android:layout_width="match_parent"
   android:layout_height="300dp">
-
 
   <ImageButton
     android:id="@+id/btn_start_next"
@@ -14,8 +13,8 @@
     android:layout_height="wrap_content"
     android:layout_marginEnd="8dp"
     android:layout_marginStart="8dp"
-    android:layout_marginTop="4dp"
     app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintHorizontal_bias="0.957"
     app:layout_constraintStart_toEndOf="@+id/textView_start_message"
     app:layout_constraintTop_toTopOf="parent"
     app:srcCompat="@android:drawable/alert_dark_frame" />
