+++ app/src/main/res/layout/item_tendency_view_flipper2.xml
@@ -1,18 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  android:background="#a97070"
-  android:orientation="vertical"
-  android:layout_margin="30dp"
-
+<android.support.constraint.ConstraintLayout
+  xmlns:android="http://schemas.android.com/apk/res/android"
   android:layout_width="match_parent"
-  android:layout_height="300dp">
+  android:layout_height="300dp"
+  android:background="#ffc2c2"
+  android:orientation="vertical">
 
   <TextView
-    android:text="두번째 View Filpper"
-    android:textSize="30dp"
     android:layout_width="wrap_content"
-    android:layout_height="wrap_content"/>
+    android:layout_height="wrap_content"
+    android:text="두번째 View Filpper"
+    android:textSize="30dp"/>
 
 </android.support.constraint.ConstraintLayout>
