+++ app/src/main/res/layout/item_tendency_view_flipper3.xml
@@ -1,17 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  android:background="#00ff19"
-  android:orientation="vertical"
+<android.support.constraint.ConstraintLayout
+  xmlns:android="http://schemas.android.com/apk/res/android"
   android:layout_width="match_parent"
   android:layout_height="300dp"
-  android:layout_margin="30dp">
+  android:background="#9fcea3"
+  android:orientation="vertical">
 
   <TextView
-    android:text="세번째 View Filpper"
-    android:textSize="30dp"
     android:layout_width="wrap_content"
-    android:layout_height="wrap_content"/>
+    android:layout_height="wrap_content"
+    android:text="세번째 View Filpper"
+    android:textSize="30dp"/>
 
 </android.support.constraint.ConstraintLayout>
